fix function types

+++ libs/blas/c/matrix.cc
@@ -2,7 +2,7 @@
 
 using Eigen::MatrixXd;
 
-extern "C" void *alloc_matrix(unsigned int rows, unsigned int cols)
+extern "C" void *alloc_matrix(unsigned long rows, unsigned long cols)
 {
     MatrixXd *m = new MatrixXd(rows, cols);
     return (void *)m;
@@ -14,13 +14,13 @@ extern "C" void free_matrix(void *m)
     delete mat;
 }
 
-extern "C" void set_matrix_val(void *m, unsigned int rows, unsigned int cols, double val)
+extern "C" void set_matrix_val(void *m, unsigned long rows, unsigned long cols, double val)
 {
     MatrixXd *mat = (MatrixXd *)m;
     (*mat)(rows, cols) = val;
 }
 
-extern "C" double get_matrix_val(void *m, unsigned int rows, unsigned int cols)
+extern "C" double get_matrix_val(void *m, unsigned long rows, unsigned long cols)
 {
     MatrixXd *mat = (MatrixXd *)m;
     return (*mat)(rows, cols);
